Added sample and docs on how to export to prometheus

+++ src/FlowtideDotNet.Base/Vertices/Egress/EgressVertex.cs
@@ -128,7 +128,7 @@ namespace FlowtideDotNet.Base.Vertices.Egress
             Metrics.CreateObservableGauge("busy", () =>
             {
                 Debug.Assert(_targetBlock != null, nameof(_targetBlock));
-                return ((decimal)_targetBlock.InputQueue) / _targetBlock.MaxInputQueue;
+                return ((float)_targetBlock.InputQueue) / _targetBlock.MaxInputQueue;
             });
 
             Metrics.CreateObservableGauge("InputQueue", () =>
